Display the logged in player on the main page.
This implements the get_player_by_id for Database and Game. For the FlatFileDatabase, this has to scan every .player file to look for that player's id, but this will be much easier with a real database.

Right now this is also loaded every time the page is loaded. That may become a problem down the road if there are many page loads, but I think it is fine for the plan of using more javascript to execute commands (rather than POST calls which would reload the page).