Add clusters to act as a container for sectors.
This adds cluster addition and retrieval for the Database class and some interfaces to those methods with the Game class.

Clusters are added when the game is started (and not during a big bang) so they can be added and removed (through a config file) as the sysop desires. If a cluster is named in the config but it does not exist, it will be created. For now, all clusters will have the same dimensions.

This also moves the load_config, connect_db, and (new) _load_clusters to the init method. This is to eliminate possible programming errors when initializing a Game, and to make sure everything is called in a specific order (clusters can only be loaded after the config and database are loaded).