A random empty sector is created for a player.
After login, the player is given a link to create a home sector. It currently asks for a planet name, and on submitting, the game will find a random empty sector, create that sector, and attempt to move the player there. Next step will be to create the planet and implement move_player().