Implemented action to move the player.
This adds a /c/<command>/<argument> to the site handlers list, but we may need to remove the argument to allow for multiple aruguments provided in the URL.

I also added comparison methods for Sector and Cluster since there can be many instances of these class that were loaded for the same game entity (same sector or same cluster).

The move_player() method re-confirms that the move is valid in case the player sends a URL they built themselves.