Moved Player to use Entity.
This required a lot of reworking, since with the idea of everything inheriting from entity, the player is a child of the ship, so when you issue a move command, you are moving the ship, not the player itself.

There was some changes to how the links appear for moving the ship on the web page, so the cluster is now included in there. It looks like a lot of text right now, so this will need to be worked on to get it looking cleaner.

This also adds load_object in Game, which will be expanded, and likely added to Database as well.