Resolved issues in creating a new player.
This changes the flow of creating a new player, after the player is created, then the ship will be created, then the planet/home sector. This is because the ship now determines where the player is, not the player object itself.

I also removed the cluster and sector from the player file for the same reasons as above.