Implemented Entity.to_dict().
This returns a dictionary for every instance variable in the class, and converts the parent and child objects to only return their ids. (If it returned the children objects, then the Cluster.to_dict() would return every object in the game)

I removed the Player.to_dict() since it is less advanced than the Entity.to_dict() function, and I tested the inherited Player.to_dict() and it worked as expected.