A shamefully monstrous commit.
This contains a ton of fixes as made light by adding the testing functions. It vastly simplifies the database operations by removing the object specific get and add functions (get_sector, add_sector) and use load_object and save_object instead.

There are also a bunch of fixes to the game, player, and cluster objects, with more to come to others as more tests are written.