Simplify the init method of Entity objects.
This uses an initial_state dictionary that defines any fields that need to be set for this entity. This will change the way these objects are created, but will simplify the creation of objects that are loaded from the database.

Currently only the Entity, Cluster, Sector, and Player objects are modified and tested, more to come later.

This also fixes the big bang by adding a bigbang parameter to the Game.__init__ method. Before, the game would try to initialize the clusters before the big bang could be called.