Convert Planet and Ship to new init method.
Now all current entity types can be loaded from the _read_json method

This also adds functions to the Game object to get parent and get children objects to load from the database (or later from the cache), as well as assigning a child object to a parent.

Using these new methods, this modifies the get player test to make sure the ship is created and child objects are property assigned (player is a child of ship, ship is a child of sector).