Got the web site working again.
Removed Game.build_ship because it is unused.
Fixed the AddHandler.post method to call updated Game methods.
Moved ship name to be a part of creating a home sector in AddHandler.
Removed /add/ship support, since it is not required anymore.
Added Game.load_object_by_id.
Changed /c/move to provide a sector id which will be more easily loaded (by id rather than cluster-sector names).
Modified CommandHandler to correctly accept the move command.
Fixed Entity.__init__ to correctly handle when self.parent is loaded from the json file as "None", it now converts to a true python None (which fixes the assign_home_sector, which skipped creating a home sector if the player had a parent which was not None).