Can now land on and takeoff from planets.
This does a lot of housekeeping as well, moving some parts of the html template to their own file (specifically the entity-panel.html), which should help prevent duplicated code later.

For now, there is a separate file for rendering a planet (planet.html), but this will likely be combined at some point to a common render.html file.

Closes #8